@@ -1,14 +1,12 @@
-import Vue from "vue";
+import { createApp } from "vue";
 import App from "./App.vue";
 import router from "./router";
 import store from "./store";
 import { preload } from "./utils/help.js";
 
-Vue.config.productionTip = false;
 preload();
 
-new Vue({
-  router,
-  store,
-  render: h => h(App)
-}).$mount("#app");
+const app = createApp(App);
+app.use(router);
+app.use(store);
+app.mount("#app");
