@@ -271,51 +271,51 @@ export default class GameStateManager {
     return {
       // 基础状态标志
       hitflag() {
-        const manager = this.gameStateManager || this.gameManager;
+        const manager = this.gameManager;
         return manager ? manager.hitflag : true;
       },
       
       lockflag() {
-        const manager = this.gameStateManager || this.gameManager;
+        const manager = this.gameManager;
         return manager ? manager.lockflag : true;
       },
       
       winflag() {
-        const manager = this.gameStateManager || this.gameManager;
+        const manager = this.gameManager;
         return manager ? manager.winflag : false;
       },
       
       loseflag() {
-        const manager = this.gameStateManager || this.gameManager;
+        const manager = this.gameManager;
         return manager ? manager.loseflag : false;
       },
       
       drawflag() {
-        const manager = this.gameStateManager || this.gameManager;
+        const manager = this.gameManager;
         return manager ? manager.drawflag : false;
       },
       
       // 步数计算属性
       step() {
-        const manager = this.gameStateManager || this.gameManager;
+        const manager = this.gameManager;
         return manager ? manager.getStepCount() : 0;
       },
       
       // 按钮禁用状态计算属性
       undoDisabled() {
-        const manager = this.gameStateManager || this.gameManager;
+        const manager = this.gameManager;
         return !manager || !manager.canUndo();
       },
       
       restartDisabled() {
-        const manager = this.gameStateManager || this.gameManager;
+        const manager = this.gameManager;
         return !manager || 
                !manager.hitflag || 
                !manager.lockflag;
       },
       
       stepDisabled() {
-        const manager = this.gameStateManager || this.gameManager;
+        const manager = this.gameManager;
         return !manager ||
                !manager.hitflag ||
                !manager.lockflag ||
@@ -325,7 +325,7 @@ export default class GameStateManager {
       },
       
       autoDisabled() {
-        const manager = this.gameStateManager || this.gameManager;
+        const manager = this.gameManager;
         return !manager ||
                !manager.hitflag ||
                !manager.lockflag ||
@@ -337,7 +337,7 @@ export default class GameStateManager {
       
       // 游戏状态
       gameState() {
-        const manager = this.gameStateManager || this.gameManager;
+        const manager = this.gameManager;
         return manager ? manager.getState() : {};
       }
     };
