@@ -1,3 +1,6 @@
-module.exports = {
-  publicPath: "./"
-};
+const { defineConfig } = require('@vue/cli-service')
+
+module.exports = defineConfig({
+  publicPath: "./",
+  transpileDependencies: true
+});
